+++ packages/wallet-coinbase/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vleap/warps-wallet-coinbase
 
+## 1.0.0-beta.2
+
+### Minor Changes
+
+- bbfdafb: add coinbase wallet
+- 3e86317: add wallet import + export
+
+### Patch Changes
+
+- 600b777: update wallets
+- Updated dependencies [f67aea4]
+- Updated dependencies [3e86317]
+  - @vleap/warps@3.0.0-beta.184
+
 ## 1.0.0-beta.1
 
 ### Minor Changes